Expand player state to 64-bit integer
Changed player state variables and related methods from uint32_t to uint64_t to support additional state flags for newer client versions. Also updated protocol parsing and getter/setter signatures accordingly. Minor fix in UIWidget HTML layout to use std::max<int> for clarity.

+++ src/client/protocolgameparse.cpp
@@ -2493,7 +2493,7 @@ void ProtocolGame::parsePlayerSkills(const InputMessagePtr& msg) const
 
 void ProtocolGame::parsePlayerState(const InputMessagePtr& msg) const
 {
-    uint32_t states;
+    uint64_t states;
     if (g_game.getClientVersion() >= 1281) {
         states = g_game.getClientVersion() >= 1405 ? msg->getU64() : msg->getU32();
         if (g_game.getFeature(Otc::GamePlayerStateCounter)) {
